test(telegram): split provider session turn coverage
